feat: add publish workflow to pub.dev
This commit adds a GitHub Actions workflow to automatically publish the package to pub.dev when a new tag matching the `v[0-9]+.[0-9]+.[0-9]+*` pattern is pushed.

The workflow uses the `dart-lang/setup-dart/.github/workflows/publish.yml@v1` action to handle the publishing process. It also requests write permission for the id-token.