feat: add toString method to enum classes for value representation
This commit adds a `toString` method to several enum classes within the project. This method overrides the default `toString` behavior to return the `value` field of the enum, which is used for serialization.

The following enum classes were updated:

- `WindowType`
- `ProductType`
- `Country`
- `PaymentLocale`
- `PGCompany`
- `PaymentPayMethod`
- `CashReceiptTradeOption`
- `SndQpayType`
- `NiceV2SkinType`
- `SmartroV2SkinColor`
- `SmartroV2OpenType`

This change improves the readability and debuggability of the code when these enums are converted to strings.