feat: ✨ add temporary ignore for avoid_catches_without_on_clauses lint
This commit adds a temporary ignore for the `avoid_catches_without_on_clauses` lint in `lib/src/portone_payment.dart`.

This is a temporary measure to allow for a more gradual refactoring of the codebase to address these lint warnings. A TODO comment has been added to ensure that this ignore is removed once the underlying issues have been resolved.